Add reporting websocket for button events
This sends the JSON format that the accessor currently expects.

It hard codes one "row" of UI for the moment, which may be enough
to just copy/paste for this demo.

The "status" bar is more naive than its intended future. The idea
would be to listen to the websocket for feedback from the robot
(e.g. 3m away... 1m away... arriving...)